% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{add_days_of_week_columns}
\alias{add_days_of_week_columns}
\title{Add columns for days of the week to smoothed data}
\usage{
add_days_of_week_columns(smoothed_cbc_features)
}
\arguments{
\item{smoothed_cbc_features}{with the date and features}
}
\value{
the data with days of the week columns added as a tibble
}
\description{
Add columns for days of the week to smoothed data
}
