% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{analyze_prediction_table}
\alias{analyze_prediction_table}
\title{Combine projection loss and prediction error calculation into a single evaluation.}
\usage{
analyze_prediction_table(pred_table, config)
}
\arguments{
\item{pred_table, }{the prediction table generated from build_prediction_table}

\item{config}{the site-specific configuration}
}
\value{
a list containing a summary of the projection loss and prediction error (i.e.
        the overall efficacy of the model)
}
\description{
Combine projection loss and prediction error calculation into a single evaluation.
}
