% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{build_prediction_table_db}
\alias{build_prediction_table_db}
\title{Database version of build_prediction_table}
\usage{
build_prediction_table_db(
  conn,
  config,
  start_date,
  end_date = Sys.Date() + 2,
  pred_tbl = NULL,
  offset = config$start - 1,
  min_inventory = config$min_inventory
)
}
\arguments{
\item{conn}{the database connection obejct}

\item{config}{the site configuration}

\item{start_date}{the starting date in YYYY-mm-dd format}

\item{end_date}{the end date in YYYY-mm-dd format, by default today plus 2 days}

\item{pred_tbl}{a table of predictions generated by predict_for_date}

\item{min_inventory}{the minimum that needs to be in inventory,
by default what was used in the training, which is config$c0.}
}
\value{
a tibble of several variables, including all columns of
    prediction_df, plus units expiring in a day (r1), units
    expiring in 2 days (r2), waste (w), collection units (x),
    shortage (s) and y prediction and the platelet usage for that
    date, suggested values in case of inventory minimum is not met,
    inventory columns if available
}
\description{
Database version of build_prediction_table
}
