% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{compute_real_loss}
\alias{compute_real_loss}
\title{Compute the Proxy Loss}
\usage{
compute_real_loss(w, r1, r2, penalty_factor, c0)
}
\arguments{
\item{w}{a vector of waste determined by pip::compute_prediction_statistics}

\item{r1}{a vector of remaining inventory determined by pip::compute_prediction_statistics}

\item{r2}{a vector of remaining inventory determined by pip::compute_prediction_statistics}

\item{penalty_factor}{factor to additionally penalize shortage terms over waste}

\item{c0}{the minimum inventory level below which we penalize.}
}
\value{
a vector of losses computed for each hyperparameter
}
\description{
Since we cannot observe shortage, or the true predicted response, the proxy 
shortage refers to the total number of inventory platelets below c0
}
