% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{create_cbc_features}
\alias{create_cbc_features}
\title{Construct a tibble containing the quartiles of the CBC values and lagged features}
\usage{
create_cbc_features(cbc, cbc_quantiles)
}
\arguments{
\item{cbc}{the cbc data tibble}

\item{cbc_quantiles}{a named list of site-specific quantile functions for each cbc}
}
\value{
a tibble of containing cbc variables of interest grouped by
    date
}
\description{
Construct a tibble containing the quartiles of the CBC values and lagged features
}
