% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{create_dataset}
\alias{create_dataset}
\title{Construct a dataset for use in forecasting}
\usage{
create_dataset(config, cbc_features, census, surgery, transfusion)
}
\arguments{
\item{config}{a site-specific configuration list}

\item{cbc_features}{the tibble of cbc features}

\item{census}{the census data}

\item{surgery}{the surgery data}

\item{transfusion}{the transfusion data}
}
\value{
a single tibble that contains the response, features and date (1st col)
}
\description{
Construct a dataset for use in forecasting
}
