% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{generate_full_dataset}
\alias{generate_full_dataset}
\title{Generate the full dataset used for prediction without relying on output RDS files}
\usage{
generate_full_dataset(pred_start_date, num_days, config)
}
\arguments{
\item{pred_start_date, }{the start date for the prediction, used to specify
date range for the seed dataset.}

\item{num_days, }{the number of days to predict from pred_start_date}

\item{config}{the site-specific configuration}
}
\value{
a dataset tibble with date, product usage (transfusion), CBC, census, and inventory
}
\description{
Useful for applying different prediction techniques to the same data. Note that
this has the side-effect of producing reports in the report folder for each date due
to the process_data_for_date function.
}
