% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{get_coefs_over_time}
\alias{get_coefs_over_time}
\title{Fetch model coefficients based on generated output files}
\usage{
get_coefs_over_time(pred_start_date, num_days, config)
}
\arguments{
\item{pred_start_date, }{the start date for the prediction, used to specify
date range for the seed dataset.}

\item{num_days, }{the number of days to predict from pred_start_date}

\item{config}{the site-specific configuration}
}
\value{
a data frame containing the coefficients generated during the prediction
}
\description{
Fetch model coefficients based on generated output files
}
