% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{get_prediction_and_usage}
\alias{get_prediction_and_usage}
\title{Get the actual prediction and platelet usage data from saved files for each date}
\usage{
get_prediction_and_usage(config, start_date, end_date)
}
\arguments{
\item{config}{the site configuration}

\item{start_date}{the starting date in YYYY-mm-dd format}

\item{end_date}{the end date in YYYY-mm-dd format}
}
\value{
a tibble of three variables: date, the corresponding
    prediction and the platelet usage for that date
}
\description{
This function reads a saved dataset and returns a tibble with a
date, platelet usage, and three day predicted sum ensuring that the
prediction and dates are lined up correctly.
}
