% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_globals.R
\name{get_SBC_features}
\alias{get_SBC_features}
\title{Obtain the exact features used in the model}
\usage{
get_SBC_features()
}
\value{
a vector of strings corresponding to database column headers/features
}
\description{
Obtain the exact features used in the model
}
