% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ma}
\alias{ma}
\title{Computing moving average of all past values including last value
using a window size}
\usage{
ma(z, window_size = 5L)
}
\arguments{
\item{z}{the numeric vector}

\item{window_size}{the integer window size}
}
\value{
a vector of the same length pre-padded with NAs
}
\description{
Computing moving average of all past values including last value
using a window size
}
