% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{model_config_changed}
\alias{model_config_changed}
\title{Check if model parameters have changed in two configurations.}
\usage{
model_config_changed(config1, config2)
}
\arguments{
\item{config1}{first list}

\item{config2}{second list}
}
\value{
TRUE or FALSE
}
\description{
We only consider those parameters that can be changed
    by the webapp in detecting change. This is used to decide if
    model needs to be rebuilt
}
