% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ms}
\alias{ms}
\title{Computing moving sum of all past values including last value
using a window size}
\usage{
ms(z, window_size = 3L)
}
\arguments{
\item{z}{the numeric vector}

\item{window_size}{the integer window size}
}
\value{
a vector of the same length pre-padded with NAs
}
\description{
Computing moving sum of all past values including last value
using a window size
}
