% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{plot_pred_vs_true_usage}
\alias{plot_pred_vs_true_usage}
\title{Plot the predicted vs. true product usage over the appropriate date range}
\usage{
plot_pred_vs_true_usage(pred_table)
}
\arguments{
\item{pred_table, }{the prediction table generated from SBCpip::build_prediction_table}
}
\description{
Plot the predicted vs. true product usage over the appropriate date range
}
