% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{pred_table_analysis}
\alias{pred_table_analysis}
\title{Combine projection loss and prediction error calculation into a single evaluation.}
\usage{
pred_table_analysis(pred_table, config)
}
\arguments{
\item{pred_table, }{the prediction table generated from SBCpip::build_prediction_table}

\item{config}{the site-specific configuration}
}
\value{
a list containing a summary of the projection loss and prediction error (i.e.
        the overall efficacy of the model)
}
\description{
Combine projection loss and prediction error calculation into a single evaluation.
}
