% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{predict_for_date}
\alias{predict_for_date}
\title{Predict usage for a specified date}
\usage{
predict_for_date(config, date = as.character(Sys.Date(), format =
  "\%Y-\%m-\%d"), prev_day = NA)
}
\arguments{
\item{config}{the site configuration}

\item{date}{the date string for which the data is to be processed in "YYYY-mm-dd" format}

\item{prev_day}{the previous date, default NA, which means it is computed from date}
}
\value{
a prediction tibble named prediction_df with a column for date and the prediction
}
\description{
This function updates the saved datasets (therefore, has
side-effects) by reading incremental data for a specified date. The
\code{prev_day} argument can be specified in case the pipeline
fails for some reason to catch up. Note that the default set up is
one where the prediction is made on the morning of day \eqn{i + 1}
for day \eqn{i}.
}
