% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{prediction_error}
\alias{prediction_error}
\title{Compute the RMSE of predicted vs. actual next-3-day product usage by day of week}
\usage{
prediction_error(pred_table, config)
}
\arguments{
\item{pred_table, }{the prediction table generated from SBCpip::build_prediction_table}

\item{config}{the site-specific configuration}
}
\value{
a list containing the RMSE for next-3-day product usage by day of week
}
\description{
Note: Even though wastage minimization is the end-goal of the model (and not just
usage prediction), this will help understand the effects of varying config hyperparameters
such as c0 and penalty_factor. We can also explore the effects of varying the
day(s) of week on which the model is updated, as well as the model_update_frequency.
}
