% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{process_all_cbc_files}
\alias{process_all_cbc_files}
\title{Process all cbc files in a folder and generate qc reports}
\usage{
process_all_cbc_files(data_folder,
  report_folder = file.path(dirname(data_folder),
  paste0(basename(data_folder), "_Reports")), cbc_abnormals, cbc_vars,
  pattern = "LAB-BB-CSRP-CBC*")
}
\arguments{
\item{data_folder}{the folder containing the raw data}

\item{report_folder}{the folder to write reports to, default is
data_folder with "_Reports" appended. Must exist.}

\item{cbc_abnormals, }{a named list of functions of a single vector
returning TRUE for abnormal values}

\item{cbc_vars, }{the names of fields to include; others are excluded}

\item{pattern}{the pattern to distinguish CBC files, default
"LAB-BB-CSRP-CBC*" appearing anywhere}
}
\value{
a combined dataset
}
\description{
Process all cbc files in a folder and generate qc reports
}
