% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_all_census_files}
\alias{process_all_census_files}
\title{Process all census files in a folder and generate qc reports}
\usage{
process_all_census_files(
  data_folder,
  report_folder = file.path(dirname(data_folder), paste0(basename(data_folder),
    "_Reports")),
  locations,
  pattern = "LAB-BB-CSRP-Census*"
)
}
\arguments{
\item{data_folder}{the folder containing the raw data}

\item{report_folder}{the folder to write reports to, default is
data_folder with "_Reports" appended. Must exist.}

\item{locations}{the character vector of locations to consider}

\item{pattern}{the pattern to distinguish CBC files, default
"LAB-BB-CSRP-Census*" appearing anywhere}
}
\value{
a combined dataset
}
\description{
Process all census files in a folder and generate qc reports
}
