% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_all_inventory_files}
\alias{process_all_inventory_files}
\title{Process all inventory files in a folder and generate qc reports}
\usage{
process_all_inventory_files(
  data_folder,
  report_folder = file.path(dirname(data_folder), paste0(basename(data_folder),
    "_Reports")),
  pattern = "Daily_Product_Inventory_Report_Morning_To_Folder*",
  org_cols = c("Inv. ID", "Type", "Days to Expire", "Exp. Date", "Exp. Time")
)
}
\arguments{
\item{data_folder}{the folder containing the raw data}

\item{report_folder}{the folder to write reports to, default is
data_folder with "_Reports" appended. Must exist.}

\item{pattern}{the pattern to distinguish CBC files, default
"Daily_Product_Inventory_Report_Morning_To_Folder*" appearing anywhere}

\item{org_cols}{a vector of organization-specific column headers for relevant inventory fields}
}
\value{
a combined dataset
}
\description{
Process all inventory files in a folder and generate qc reports
}
