% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{process_data_for_date}
\alias{process_data_for_date}
\title{Process data for a particular date by reading files per configuration and date}
\usage{
process_data_for_date(
  config,
  date = as.character(Sys.Date(), format = "\%Y-\%m-\%d")
)
}
\arguments{
\item{config}{a site-specific configuration list}

\item{date}{a string in YYYY-mm-dd format, default today (Day i + 1 vs. prediction for i + 1 - i + 3)}
}
\value{
a list of cbc, census, and transfusion tibbles
}
\description{
This has a side-effect of producing reports in the reports folder
specified in the site-specific configuration object as per
\code{\link{get_SBC_config}}
}
