% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{projection_loss}
\alias{projection_loss}
\title{Compute the loss value on a validation or test dataset from the original
and adjusted waste and inventory levels based on model projections.}
\usage{
projection_loss(pred_table, config)
}
\arguments{
\item{pred_table, }{the prediction table generated from SBCpip::build_prediction_table}

\item{config}{the site-specific configuration}
}
\value{
a list containing the projected average daily loss and adjusted average daily loss
}
\description{
The loss formula is given by the sum of the waste plus the sum of the squared positive
difference between the desired inventory level and the remaining inventory, as per
pip::build_model. This can be used to standardize and compare the efficacy of different
models.
}
