% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{read_next_three_surgery_files}
\alias{read_next_three_surgery_files}
\title{Read a surgery file data  for following 3 days and return tibble and summary}
\usage{
read_next_three_surgery_files(
  filenames,
  services,
  org_cols = c("LOG_ID", "OR_SERVICE", "SURGERY_DATE", "FIRST_SCHED_DATE",
    "CASE_CLASS")
)
}
\arguments{
\item{filenames}{the fully qualified path of the 3 files file}

\item{services}{the list of surgery types considered as features}
}
\value{
a list of four items, filename, raw_data (tibble), report a
    list consisting of summary tibble, surgery_data (tibble)
}
\description{
Read a surgery file data  for following 3 days and return tibble and summary
}
