% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{read_one_transfusion_file}
\alias{read_one_transfusion_file}
\title{Read a single transfusion file data and return it, as is}
\usage{
read_one_transfusion_file(filename)
}
\arguments{
\item{filename}{the fully qualified path of the file}
}
\value{
a list of four items, filename, raw_data (tibble), report a
    list consisting of summary tibble, census_data (tibble)
}
\description{
Read a single transfusion file data and return it, as is
}
