% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{real_loss}
\alias{real_loss}
\title{Compute the true average daily loss based on actual inventory levels during the
prediction period.}
\usage{
real_loss(pred_table, config)
}
\arguments{
\item{pred_table, }{the prediction table generated from SBCpip::build_prediction_table}

\item{config}{the site-specific configuration}
}
\value{
a list containing the true average daily loss
}
\description{
The loss formula is given by the sum of the waste plus the sum of the squared positive
difference between the desired inventory level and the remaining inventory, as per
pip::build_model. This can be used to standardize and compare the efficacy of different
models.
}
