% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{recommend_collection}
\alias{recommend_collection}
\title{Generate a recommendation for number of platelets to collect today based on 
predicted usage.}
\usage{
recommend_collection(
  conn,
  config,
  date,
  pred,
  product_type = 2L,
  x_plus1 = 30,
  x_plus2 = 30,
  r_1 = 0,
  r_2 = 0
)
}
\arguments{
\item{conn}{the active database connection object}

\item{config}{the configuration object}

\item{date}{the date on which we expect to collect new platelets}

\item{pred}{predicted total platelet usage over the next [product_type + 1] days}

\item{product_type}{either 1 or 2 depending on the number of days taken to 
prepare the blood product.}

\item{x_plus1}{the number of platelets collected on day [date - product_type], 
which will be available on day [date]}

\item{x_plus2}{the number of platelets collected on day [date + 1 - product_type],
which will be available on day [date + 1]. Irrelevant if product_type == 1}

\item{r_1}{the number of platelets considered expired after [date]}

\item{r_2}{the number of platelets considered expired after [date + 1]}
}
\value{
recommended number of platelets to collect on [date]
}
\description{
Generate a recommendation for number of platelets to collect today based on 
predicted usage.
}
