% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_globals.R
\name{reset_config}
\alias{reset_config}
\title{Reset the global package variable \code{sbc_config}}
\usage{
reset_config()
}
\value{
the default value value of testpack package global \code{sbc_config}
}
\description{
Reset the global package variable \code{sbc_config}
}
\examples{
\dontrun{
  reset_config()
}

}
