% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{save_report_file}
\alias{save_report_file}
\title{Save a generated report file in the report folder with the appropriate name}
\usage{
save_report_file(report_tbl, report_folder, filename)
}
\arguments{
\item{report_tbl}{the report tibble}

\item{report_folder}{the report folder}

\item{filename}{the name of the file}
}
\description{
Save a generated report file in the report folder with the appropriate name
}
