% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{sbc_build_and_save_full_db}
\alias{sbc_build_and_save_full_db}
\title{Preprocess all available files and add to database}
\usage{
sbc_build_and_save_full_db(conn, config, updateProgress = NULL)
}
\arguments{
\item{conn}{the active database connection object}

\item{config}{the site-specific configuration}

\item{updateProgress}{the callback function to display progress in the shiny app}
}
\value{
nothing (saves an RDS file with the seed data for 1 day before prediction date)
}
\description{
Preprocess all available files and add to database
}
