% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{sbc_build_and_save_seed_data}
\alias{sbc_build_and_save_seed_data}
\title{Generate .RDS file for seed data to be used in prediction}
\usage{
sbc_build_and_save_seed_data(pred_start_date, config)
}
\arguments{
\item{pred_start_date, }{the start date for the prediction, used to specify
date range for the seed dataset.}

\item{config}{the site-specific configuration}
}
\value{
nothing (saves an RDS file with the seed data for 1 day before prediction date)
}
\description{
Generate .RDS file for seed data to be used in prediction
}
