% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{sbc_build_and_save_seed_db}
\alias{sbc_build_and_save_seed_db}
\title{Generate seed database for use in prediction}
\usage{
sbc_build_and_save_seed_db(conn, pred_start_date, config)
}
\arguments{
\item{conn}{the active database connection object}

\item{pred_start_date, }{the start date for the prediction, used to specify
date range for the seed dataset.}

\item{config}{the site-specific configuration}
}
\value{
nothing (saves an RDS file with the seed data for 1 day before prediction date)
}
\description{
Generate seed database for use in prediction
}
