% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_wrapper.R
\name{sbc_predict_for_range_db}
\alias{sbc_predict_for_range_db}
\title{Database version of range prediction}
\usage{
sbc_predict_for_range_db(
  conn,
  pred_start_date,
  num_days,
  config,
  updateProgress = NULL
)
}
\arguments{
\item{conn, }{the database connection object}

\item{pred_start_date, }{the start date for the prediction, used to specify
date range for the seed dataset.}

\item{num_days, }{the number of days to predict from pred_start_date}

\item{config}{the site-specific configuration}

\item{updateProgress}{a callback function to update progress for the shiny app (optional)}
}
\value{
a prediction tibble named prediction_df with a column for date and the prediction.
    This function is essentially a wrapper for SBCpip::predict_for_date across a specific
    date range.
}
\description{
Database version of range prediction
}
\note{
Assumes that the config output folder contains a seed dataset labeled one day prior
    to the prediction date range.
}
