% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{scale_dataset}
\alias{scale_dataset}
\title{Scale the dataset and return the scaling parameters}
\usage{
scale_dataset(dataset, center = NULL, scale = NULL)
}
\arguments{
\item{dataset}{the dataset of date column, plus features}

\item{center}{the values to use for centering if not NULL}

\item{scale}{the values to use for scaling if not NULL}
}
\value{
a list consisting of scaled data (date not touched) and center and scale
}
\description{
Scale the dataset and return the scaling parameters
}
