% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_globals.R
\name{set_config_param}
\alias{set_config_param}
\title{Set SBC configuration parameter to a specified value}
\usage{
set_config_param(param, value)
}
\arguments{
\item{param}{the name of the variable as a string}

\item{value}{the value to assign}
}
\value{
the changed value of the package global \code{sbc_config} invisibly
}
\description{
Set SBC configuration parameter to a specified value
}
