% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_globals.R
\name{set_features_from_file}
\alias{set_features_from_file}
\title{Gather and set feature names from a single (most recent) data file}
\usage{
set_features_from_file()
}
\value{
a list of the cbc, census, and surgery features derived from the data file
}
\description{
Requires config$data_folder, config$cbc_filename_prefix, 
config$census_filename_prefix, config$surgery_filename_prefix to be valid.
}
