% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_globals.R
\name{set_initial_expiry_data}
\alias{set_initial_expiry_data}
\title{Set the SBC initial number of units that will expire in a day and
the next}
\usage{
set_initial_expiry_data(value)
}
\arguments{
\item{value}{the value to assign, should be a non-negative 2-vector}
}
\value{
the changed value of the package global \code{sbc_config}
    invisibly
}
\description{
Set the SBC initial number of units that will expire in a day and
the next
}
