% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_globals.R
\name{set_org_col_params}
\alias{set_org_col_params}
\title{Set the important column names contained in the organization's file types}
\usage{
set_org_col_params()
}
\arguments{
\item{file_type}{one of c("cbc", "census", "surgery", "transfusion", "inventory")}
}
\value{
the changed value of the package global \code{sbc_config}
    invisibly
}
\description{
Set the important column names contained in the organization's file types
}
