% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{smooth_cbc_features}
\alias{smooth_cbc_features}
\title{Smooth the CBC features using a moving average}
\usage{
smooth_cbc_features(cbc_data, window_size = 7L)
}
\arguments{
\item{cbc_data}{with the date and features}

\item{window_size}{the window size to use in smoothing, default 7}
}
\value{
the data with features smoothed as a tibble
}
\description{
Smooth the CBC features using a moving average
}
