% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{summarize_and_clean_cbc}
\alias{summarize_and_clean_cbc}
\title{Summarize and clean the raw cbc data}
\usage{
summarize_and_clean_cbc(raw_data, cbc_abnormals, cbc_vars)
}
\arguments{
\item{raw_data}{the raw data tibble}

\item{cbc_abnormals, }{a named list of functions of a single vector
returning TRUE for abnormal values}

\item{cbc_vars, }{the names of fields to include; others are excluded}
}
\value{
a list of five items; errorCode (nonzero if error),
    errorMessage if any, missing_value_df the missing value data
    tibble, the summary data tibble, the data tibble filtered with
    relevant columns for us
}
\description{
Summarize and clean the raw cbc data
}
