% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{summarize_and_clean_census}
\alias{summarize_and_clean_census}
\title{Summarize and clean the raw census data}
\usage{
summarize_and_clean_census(raw_data, locations)
}
\arguments{
\item{raw_data}{the raw data tibble}

\item{locations}{a character vector locations of interest}
}
\value{
a list of five items; errorCode (nonzero if error),
    errorMessage if any, missing_value_df the missing value data
    tibble, the summary data tibble, the data tibble filtered with
    relevant columns for us
}
\description{
Summarize and clean the raw census data
}
