% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{summarize_and_clean_inventory}
\alias{summarize_and_clean_inventory}
\title{Summarize and clean the raw transfusion data}
\usage{
summarize_and_clean_inventory(raw_data, date)
}
\arguments{
\item{raw_data}{the raw data tibble}

\item{date}{the date for which this is the inventory}
}
\value{
a list of four items; errorCode (nonzero if error),
    errorMessage if any, the summary data tibble, the data tibble
    filtered with relevant columns for us
}
\description{
Summarize and clean the raw transfusion data
}
