% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{summarize_and_clean_surgery_single_day}
\alias{summarize_and_clean_surgery_single_day}
\title{Summarize and clean the raw surgery data}
\usage{
summarize_and_clean_surgery_single_day(raw_data, services)
}
\arguments{
\item{raw_data}{the raw data tibble}

\item{services}{the list of surgery types considered as features}
}
\value{
a list of four items; errorCode (nonzero if error),
    errorMessage if any, the summary data tibble, the data tibble
    filtered with relevant columns for us
}
\description{
Summarize and clean the raw surgery data
}
