% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{summarize_and_clean_transfusion}
\alias{summarize_and_clean_transfusion}
\title{Summarize and clean the raw transfusion data}
\usage{
summarize_and_clean_transfusion(raw_data)
}
\arguments{
\item{raw_data}{the raw data tibble}
}
\value{
a list of four items; errorCode (nonzero if error),
    errorMessage if any, the summary data tibble, the data tibble
    filtered with relevant columns for us
}
\description{
Summarize and clean the raw transfusion data
}
