% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{upsert_to_database}
\alias{upsert_to_database}
\title{Helper Database upsert function}
\usage{
upsert_to_database(conn, df_name, new_row, by = "date")
}
\arguments{
\item{conn}{the database connection object}

\item{df_name}{string name of the database table}

\item{new_row}{new complete row to add to the database table (data.frame)}

\item{by}{string name of column by which to match rows (default to date for this use case)}
}
\description{
Helper Database upsert function
}
