% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{read_one_inventory_file}
\alias{read_one_inventory_file}
\title{Read a single inventory file and return it, as is}
\usage{
read_one_inventory_file(filename, date = NULL)
}
\arguments{
\item{filename}{the fully qualified path of the file}

\item{date}{an optional date of inventory in case the date is not
to be automatically inferred as the day before the date in the
filename}
}
\value{
a list of four items, filename, raw_data (tibble), report a
    list consisting of summary tibble, census_data (tibble)
}
\description{
Read a single inventory file and return it, as is
}
