% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-data.R
\name{read_one_surgery_file}
\alias{read_one_surgery_file}
\title{Read a single surgery file data and return tibble and summary}
\usage{
read_one_surgery_file(filename, services)
}
\arguments{
\item{filename}{the fully qualified path of the file}

\item{services}{the list of surgery types considered as features}
}
\value{
a list of four items, filename, raw_data (tibble), report a
    list consisting of summary tibble, surgery_data (tibble)
}
\description{
Read a single surgery file data and return tibble and summary
}
