% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{read_one_transfusion_file}
\alias{read_one_transfusion_file}
\title{Read a single transfusion file data and return it, as is}
\usage{
read_one_transfusion_file(
  filename,
  org_cols = c("DIN", "Product Code", "Type", "Issue Date/Time")
)
}
\arguments{
\item{filename}{the fully qualified path of the file}

\item{org_cols}{list of target organization's column names (from data mapping)}
}
\value{
a list of four items, filename, raw_data (tibble), report a
    list consisting of summary tibble, transfusion_data (tibble)
}
\description{
Read a single transfusion file data and return it, as is
}
